requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
lxml==5.1.0
numpy==1.26.4
openai==1.12.0
opencv-python==4.9
packaging==23.2
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pytesseract==0.3.10
python-docx==1.1.0
python-dotenv==1.0.1
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
